Toevoegen monitoring van jeroen

+++ src/nerdygadgets/Monitoring/Project.java
@@ -3,9 +3,9 @@ package nerdygadgets.Monitoring;
 class Project {
     int ProjectID;
     String name;
-    int wanted_availability;
+    double wanted_availability;
 
-    Project(int projectID,String name,int wanted_availability) {
+    Project(int projectID,String name,double wanted_availability) {
         this.ProjectID = projectID;
         this.name = name;
         this.wanted_availability = wanted_availability;
